# Conflicts:
#   inv/models/vendor.py

+++ lib/app/extmodelapplication.py
@@ -2,11 +2,12 @@
 # ----------------------------------------------------------------------
 # ExtModelApplication implementation
 # ----------------------------------------------------------------------
-# Copyright (C) 2007-2017 The NOC Project
+# Copyright (C) 2007-2018 The NOC Project
 # See LICENSE for details
 # ----------------------------------------------------------------------
 
 # Python modules
+from __future__ import absolute_import
 import datetime
 from functools import reduce
 # Third-party modules
@@ -19,16 +20,18 @@ from django.db.utils import IntegrityError
 from django.core.exceptions import ValidationError
 import six
 # NOC modules
-from extapplication import ExtApplication, view
 from noc.sa.interfaces.base import (
     BooleanParameter, IntParameter,
     FloatParameter, TagsParameter,
     NoneParameter, StringListParameter,
     DictParameter, ListOfParameter,
     ModelParameter, InterfaceTypeError)
-from interfaces import DateParameter, DateTimeParameter
 from noc.lib.validators import is_int
 from noc.models import is_document
+from noc.main.models.tag import Tag
+from noc.core.stencil import stencil_registry
+from .extapplication import ExtApplication, view
+from .interfaces import DateParameter, DateTimeParameter
 
 
 class ExtModelApplication(ExtApplication):
@@ -38,9 +41,11 @@ class ExtModelApplication(ExtApplication):
     query_condition = "startswith"  # Match method for string fields
     int_query_fields = []  # Query integer fields for exact match
     pk_field_name = None  # Set by constructor
-    clean_fields = {}  # field name -> Parameter instance
+    clean_fields = {"id": IntParameter()}  # field name -> Parameter instance
     custom_fields = {}  # name -> handler, populated automatically
     order_map = {}  # field name -> SQL query for ordering
+    lookup_default = [{"id": "Leave unchanged", "label": "Leave unchanged"}]
+    ignored_fields = set(["id", "bi_id"])
 
     def __init__(self, *args, **kwargs):
         super(ExtModelApplication, self).__init__(*args, **kwargs)
@@ -76,7 +81,7 @@ class ExtModelApplication(ExtApplication):
                                  if f.unique and isinstance(f, CharField)]
         # Add searchable custom fields
         self.query_fields += ["%s__%s" % (f.name, self.query_condition)
-            for f in self.get_custom_fields() if f.is_searchable]
+                              for f in self.get_custom_fields() if f.is_searchable]
 
     def get_validator(self, field):
         """
@@ -104,7 +109,7 @@ class ExtModelApplication(ExtApplication):
         elif isinstance(field, related.ForeignKey):
             self.fk_fields[field.name] = field.rel.to
             return ModelParameter(field.rel.to,
-                required=not field.null)
+                                  required=not field.null)
         else:
             return None
 
@@ -142,8 +147,8 @@ class ExtModelApplication(ExtApplication):
                 return f
 
         q = reduce(lambda x, y: x | Q(**{get_q(y): query}),
-            self.query_fields[1:],
-            Q(**{get_q(self.query_fields[0]): query}))
+                   self.query_fields[1:],
+                   Q(**{get_q(self.query_fields[0]): query}))
         if self.int_query_fields and is_int(query):
             v = int(query)
             for f in self.int_query_fields:
@@ -167,10 +172,10 @@ class ExtModelApplication(ExtApplication):
         :return: dict of cleaned parameters of raised InterfaceTypeError
         :rtype: dict
         """
-        # Strip "id" and convert empty strings to None
+        # Strip ignored fields and convert empty strings to None
         data = dict(
-            (k, data[k] if data[k] != "" else None)
-            for k in data if k != "id"
+            (str(k), data[k] if data[k] != "" else None)
+            for k in data if k not in self.ignored_fields
         )
         # Clean up fields
         for f in self.clean_fields:
@@ -188,7 +193,7 @@ class ExtModelApplication(ExtApplication):
                     data[l] = self.clean_fields[l].clean(0)
                 del data[r]  # Dereferenced
         # Clone data
-        return dict((str(k), data[k]) for k in data)
+        return data
 
     def cleaned_query(self, q):
         nq = {}
@@ -202,10 +207,10 @@ class ExtModelApplication(ExtApplication):
             else:
                 np, lt = p, None
                 # Skip ignored params
-            if np in self.ignored_params or p in (
-                self.limit_param, self.page_param, self.start_param,
-                self.format_param, self.sort_param, self.query_param,
-                self.only_param):
+            if np in self.ignored_params or p in (self.limit_param,
+                                                  self.page_param, self.start_param,
+                                                  self.format_param, self.sort_param,
+                                                  self.query_param, self.only_param):
                 continue
             v = q[p]
             if self.in_param in p:
@@ -222,7 +227,7 @@ class ExtModelApplication(ExtApplication):
                         self.model._meta.db_table, self.model._meta.pk.name,
                         model._meta.get_field_by_name(fn)[0].attname,
                         model._meta.db_table
-                        )
+                    )
                     if None in nq:
                         nq[None] += [extra_where]
                     else:
@@ -253,6 +258,11 @@ class ExtModelApplication(ExtApplication):
             if f.name == "tags":
                 # Send tags as a list
                 r[f.name] = getattr(o, f.name)
+            elif f.name == "shape":
+                if o.shape:
+                    v = stencil_registry.get(o.shape)
+                    r[f.name] = v.id
+                    r["%s__label" % f.name] = unicode(v.title)
             elif hasattr(f, "document"):
                 # DocumentReferenceField
                 v = getattr(o, f.name)
@@ -264,9 +274,8 @@ class ExtModelApplication(ExtApplication):
                     r["%s__label" % f.name] = ""
             elif f.rel is None:
                 v = f._get_val_from_obj(o)
-                if (v is not None and
-                    type(v) not in (str, unicode, int, long, bool, list)):
-                    if type(v) == datetime.datetime:
+                if v is not None and not isinstance(v, (str, unicode, int, long, bool, list)):
+                    if isinstance(v, datetime.datetime):
                         v = v.isoformat()
                     else:
                         v = unicode(v)
@@ -354,9 +363,9 @@ class ExtModelApplication(ExtApplication):
                 fname = o[1:]
             else:
                 fname = o
-            if fname in self.order_map:
+            if o in self.order_map:
                 no = "%s_order_%d" % (fname, n)
-                extra_select[no] = self.order_map[fname]
+                extra_select[no] = self.order_map[o]
                 new_order += [no]
             else:
                 new_order += [o]
@@ -365,13 +374,46 @@ class ExtModelApplication(ExtApplication):
             extra["select"] = extra_select
         return extra, new_order
 
+    def can_create(self, user, obj):
+        """
+        Check user can create object. Used to additional
+        restrictions after permissions check
+        :param user:
+        :param obj: Object instance
+        :return: True if access granted
+        """
+        return True
+
+    def can_update(self, user, obj):
+        """
+        Check user can update object. Used to additional
+        restrictions after permissions check
+        :param user:
+        :param obj: Object instance
+        :return: True if access granted
+        """
+        return True
+
+    def can_delete(self, user, obj):
+        """
+        Check user can delete object. Used to additional
+        restrictions after permissions check
+        :param user:
+        :param obj: Object instance
+        :return: True if access granted
+        """
+        return True
+
     @view(method=["GET"], url="^$", access="read", api=True)
     def api_list(self, request):
         return self.list_data(request, self.instance_to_dict)
 
     @view(method=["GET"], url=r"^lookup/$", access="lookup", api=True)
     def api_lookup(self, request):
-        return self.list_data(request, self.instance_to_lookup)
+        try:
+            return self.list_data(request, self.instance_to_lookup)
+        except ValueError:
+            return self.response(self.lookup_default, status=self.OK)
 
     @view(method=["POST"], url="^$", access="create", api=True)
     def api_create(self, request):
@@ -380,6 +422,7 @@ class ExtModelApplication(ExtApplication):
         try:
             attrs = self.clean(attrs)
         except ValueError as e:
+            self.logger.info("Bad request: %r (%s)", request.raw_post_data, e)
             return self.render_json(
                 {
                     "success": False,
@@ -391,7 +434,7 @@ class ExtModelApplication(ExtApplication):
             # (Django raises exception on pyRules)
             # @todo: Check unique fields only?
             qattrs = dict((k, attrs[k])
-                for k in attrs if not callable(attrs[k]))
+                          for k in attrs if not callable(attrs[k]))
             # Check for duplicates
             self.queryset(request).get(**qattrs)
             return self.render_json(
@@ -404,12 +447,12 @@ class ExtModelApplication(ExtApplication):
             return self.render_json({
                 "status": False,
                 "message": "Duplicated record"
-                }, status=self.CONFLICT)
+            }, status=self.CONFLICT)
         except self.model.DoesNotExist:
             o = self.model(**attrs)
             # Run models validators
             try:
-                o.clean_fields()
+                o.full_clean(exclude=list(self.ignored_fields))
             except ValidationError as e:
                 e_msg = []
                 for f in e.message_dict:
@@ -418,6 +461,12 @@ class ExtModelApplication(ExtApplication):
                     "status": False,
                     "message": "Validation error: %s" % " | ".join(e_msg)
                 }, status=self.BAD_REQUEST)
+            # Check permissions
+            if not self.can_create(request.user, o):
+                return self.render_json({
+                    "status": False,
+                    "message": "Permission denied"
+                }, status=self.FORBIDDEN)
             # Check for duplicates
             try:
                 o.save()
@@ -452,7 +501,7 @@ class ExtModelApplication(ExtApplication):
         if only:
             only = only.split(",")
         return self.response(self.instance_to_dict(o, fields=only),
-            status=self.OK)
+                             status=self.OK)
 
     @view(method=["PUT"], url="^(?P<id>\d+)/?$", access="update", api=True)
     def api_update(self, request, id):
@@ -461,6 +510,7 @@ class ExtModelApplication(ExtApplication):
         try:
             attrs = self.clean(attrs)
         except ValueError as e:
+            self.logger.info("Bad request: %r (%s)", request.raw_post_data, e)
             return self.render_json(
                 {
                     "success": False,
@@ -479,6 +529,16 @@ class ExtModelApplication(ExtApplication):
             o = self.queryset(request).get(**{self.pk: int(id)})
         except self.model.DoesNotExist:
             return HttpResponse("", status=self.NOT_FOUND)
+        # Tags
+        if hasattr(o, "tags") and attrs.get("tags"):
+            old_tags = set(o.tags) if o.tags else set()
+            new_tags = set(attrs["tags"]) if attrs["tags"] else set()
+            for t in old_tags - new_tags:
+                self.logger.info("Unregister Tag: %s" % t)
+                Tag.unregister_tag(t, repr(self.model))
+            for t in new_tags - old_tags:
+                self.logger.info("Register Tag: %s" % t)
+                Tag.register_tag(t, repr(self.model))
         # Update attributes
         for k, v in attrs.items():
             setattr(o, k, v)
@@ -493,6 +553,12 @@ class ExtModelApplication(ExtApplication):
                 "status": False,
                 "message": "Validation error: %s" % " | ".join(e_msg)
             }, status=self.BAD_REQUEST)
+        # Check permissions
+        if not self.can_create(request.user, o):
+            return self.render_json({
+                "status": False,
+                "message": "Permission denied"
+            }, status=self.FORBIDDEN)
         # Save
         try:
             o.save()
@@ -522,6 +588,12 @@ class ExtModelApplication(ExtApplication):
                 "status": False,
                 "message": "Not found"
             }, status=self.NOT_FOUND)
+        # Check permissions
+        if not self.can_delete(request.user, o):
+            return self.render_json({
+                "status": False,
+                "message": "Permission denied"
+            }, status=self.FORBIDDEN)
         try:
             o.delete()
         except ValueError as e:
@@ -543,6 +615,7 @@ class ExtModelApplication(ExtApplication):
         try:
             v = validator.clean(rv)
         except InterfaceTypeError as e:
+            self.logger.info("Bad request: %r (%s)", request.raw_post_data, e)
             return self.render_json({
                 "status": False,
                 "message": "Bad request",
